🐛 vue-dot: Fix global types not included in package types
