--- packages/vue-dot/src/global.d.ts
@@ -1,27 +1,0 @@
-import { VueConstructor } from 'vue';
-import { VueDotOptions } from '../types';
-import { Framework } from 'vuetify/types';
-
-// Extend global interfaces
-
-/* eslint-disable @typescript-eslint/no-explicit-any */
-type UnknownValue = any;
-
-declare global {
-	interface Window {
-		Vue?: VueConstructor;
-	}
-
-	/** @see https://github.com/microsoft/TypeScript/issues/45612 */
-	interface Navigator {
-		msSaveBlob(blob: UnknownValue, defaultName?: string): boolean;
-		msSaveOrOpenBlob(blob: UnknownValue, defaultName?: string): boolean;
-	}
-}
-
-declare module 'vue/types/vue' {
-	interface Vue {
-		$vd: VueDotOptions;
-		$vuetify: Framework;
-	}
-}
