+++ packages/vue-dot/types/index.d.ts
@@ -1,4 +1,5 @@
-import { PluginFunction } from 'vue';
+import { VueConstructor, PluginFunction } from 'vue';
+import { Framework } from 'vuetify/types';
 
 // Types for the library usage
 // eg. Vue.use(VueDot, options)
@@ -27,4 +28,25 @@ export interface VueDotPlugin {
 }
 
 declare const VueDot: VueDotPlugin;
+
+declare global {
+	interface Window {
+		Vue?: VueConstructor;
+	}
+
+	/** @see https://github.com/microsoft/TypeScript/issues/45612 */
+	interface Navigator {
+		/* eslint-disable @typescript-eslint/no-explicit-any */
+		msSaveBlob(blob: any, defaultName?: string): boolean;
+		msSaveOrOpenBlob(blob: any, defaultName?: string): boolean;
+	}
+}
+
+declare module 'vue/types/vue' {
+	interface Vue {
+		$vd: VueDotOptions;
+		$vuetify: Framework;
+	}
+}
+
 export default VueDot;
